Rename
rename vestiary to locker
farming to schedule
