+++ src/pages/dashboard/dashboard.routes.jsx
@@ -1,8 +1,8 @@
 import { Routes, Route } from "react-router-dom";
 import DashboardPage from "./dashboard.page";
-import VestiaryPage from "./vestiary.page";
+import LockerPage from "./locker.page";
 import DatalabPage from "./datalab.page";
-import FarmingRoutes from "./farming/farming.routes";
+import ScheduleRoutes from "./schedule/schedule.routes";
 import PlayingRoutes from "./playing/playing.routes";
 
 export default function DashboardRoutes() {
@@ -10,9 +10,9 @@ export default function DashboardRoutes() {
     <>
       <Routes>
         <Route path="/" element={<DashboardPage />} />
-        <Route path="/vestiary" element={<VestiaryPage />} />
+        <Route path="/locker" element={<LockerPage />} />
         <Route path="/datalab" element={<DatalabPage />} />
-        <Route path="/farming/*" element={<FarmingRoutes />} />
+        <Route path="/schedule/*" element={<ScheduleRoutes />} />
         <Route path="/playing/*" element={<PlayingRoutes />} />
       </Routes>
     </>
