+++ src/pages/dashboard/schedule/schedule.routes.jsx
@@ -1,13 +1,13 @@
 import { Routes, Route } from "react-router-dom";
-import FarmingPage from "./farming.page";
 import DailyPage from "./daily.page";
 import MonthlyPage from "./monthlypage.page";
+import SchedulePage from "./schedule.page";
 
-export default function PolicyRoutes() {
+export default function ScheduleRoutes() {
   return (
     <>
       <Routes>
-        <Route path="/" element={<FarmingPage />} />
+        <Route path="/" element={<SchedulePage />} />
         <Route path="/daily" element={<DailyPage />} />
         <Route path="/monthly" element={<MonthlyPage />} />
       </Routes>
