+++ src/shared/layout/header/mobile/mobile-nav.jsx
@@ -1,46 +1,41 @@
-import { useState, useEffect } from 'react';
-import { Menu, X } from 'lucide-react';
-import MobileMenu from './mobile-menu';
-import { A } from '@img/index';
+import { useState, useEffect } from "react";
+import { Menu, X } from "lucide-react";
+import MobileMenu from "./mobile-menu";
+import { A } from "@img/index";
 
 export default function MobileNav() {
   const [isOpen, setIsOpen] = useState(false);
 
   const menuItems = [
-    { path: '/dashboard', label: 'Dashboard' },
-    { path: '/dashboard/vestiary', label: 'Vestiary' },
-    { path: '/dashboard/datalab', label: 'Datalab' },
-    { path: '/dashboard/farming', label: 'Farming' },
-    { path: '/dashboard/playing', label: 'Playing' },
+    { path: "/dashboard", label: "Dashboard" },
+    { path: "/economy", label: "Economy" },
   ];
+
   useEffect(() => {
     const handleEscape = (e) => {
-      if (e.key === 'Escape') setIsOpen(false);
+      if (e.key === "Escape") setIsOpen(false);
     };
 
-    document.addEventListener('keydown', handleEscape);
-    return () => document.removeEventListener('keydown', handleEscape);
+    document.addEventListener("keydown", handleEscape);
+    return () => document.removeEventListener("keydown", handleEscape);
   }, []);
 
   useEffect(() => {
     if (isOpen) {
-      document.body.style.overflow = 'hidden';
+      document.body.style.overflow = "hidden";
     } else {
-      document.body.style.overflow = 'unset';
+      document.body.style.overflow = "unset";
     }
 
     return () => {
-      document.body.style.overflow = 'unset';
+      document.body.style.overflow = "unset";
     };
   }, [isOpen]);
 
   return (
     <>
       <div className="h-20 flex justify-between items-center p-4">
-        <button
-          onClick={() => setIsOpen(!isOpen)}
-          className=""
-        >
+        <button onClick={() => setIsOpen(!isOpen)} className="">
           {isOpen ? <X size={24} /> : <Menu size={24} />}
         </button>
         <img src={A} alt="logo" className="h-full" />
@@ -48,8 +43,11 @@ export default function MobileNav() {
         <div>Connection</div>
       </div>
 
-      <MobileMenu menuItems={menuItems} isOpen={isOpen} onClose={() => setIsOpen(false)} />
+      <MobileMenu
+        menuItems={menuItems}
+        isOpen={isOpen}
+        onClose={() => setIsOpen(false)}
+      />
     </>
   );
 }
-
